# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "API endpoints for managing budgets",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/materials": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Get all materials",
                "operationId": "2525f22d3248fb7bb32aeb68351b2784",
                "responses": {
                    "200": {
                        "description": "List of materials retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Material"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "Create a new material",
                "operationId": "4abd9d3b4e2950611d3a2c96ca57d019",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaterialRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Material created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/materials/{id}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Get material by ID",
                "operationId": "13c74a5d8dc45d44c2eb70d5cf488285",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Material ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Material not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Materials"
                ],
                "summary": "Update an existing material",
                "operationId": "7a4879d801950933d5ae8eedfa351068",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaterialRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Material updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Materials"
                ],
                "summary": "Delete a material",
                "operationId": "45750a3aea9c3cd339f4a1d73836c164",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Material deleted successfully"
                    },
                    "404": {
                        "description": "Material not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Material": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "sub_category_id": {
                        "type": "integer"
                    },
                    "measure_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "MaterialRequest": {
                "required": [
                    "name",
                    "sub_category_id",
                    "measure_id"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "sub_category_id": {
                        "type": "integer"
                    },
                    "measure_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}